- frontends/php/httpmon.php

+++ src/zabbix_server/poller/checks_ipmi.c
@@ -1349,7 +1349,7 @@ int	parse_ipmi_command(const char *command, char *c_name, int *val, char *error,
 
 	if (ITEM_IPMI_SENSOR_LEN_MAX <= sz_c_name)
 	{
-		zbx_snprintf(error, max_error_len, "IPMI command is too long [%.*s]", sz_c_name, command);
+		zbx_snprintf(error, max_error_len, "IPMI command is too long [%.*s]", (int)sz_c_name, command);
 		goto fail;
 	}
